Factor out some helper code into util.py
