+++ python/test_util.py
@@ -1,0 +1,28 @@
+from unittest import TestCase
+
+from util import fmt_seconds, overlay_lists
+
+
+class TestUtil(TestCase):
+    def test_fmt_seconds(self):
+        self.assertEqual(fmt_seconds(1.567), '1.57')
+        self.assertEqual(fmt_seconds(59.899), '59.90')
+        self.assertEqual(fmt_seconds(59.999), '1:00.0')
+        self.assertEqual(fmt_seconds(61.789), '1:01.8')
+        self.assertEqual(fmt_seconds(3599.99), '1:00:00')
+        self.assertEqual(fmt_seconds(3601.5), '1:00:02')
+
+    def test_overlay_lists(self):
+        data = (
+            (0, ('---', '+--', '++-', '+++', '+++')),
+            (1, ('---', '-+-', '-++', '-++', '-++')),
+            (5, ('---', '---', '---', '---', '---')),
+            (-1, ('---', '---', '+--', '++-', '+++')),
+            (-4, ('---', '---', '---', '---', '---')),
+        )
+        for offset, expecteds in data:
+            for i, expected in enumerate(expecteds):
+                base = ['-'] * 3
+                new = ['+'] * i
+                overlay_lists(base, new, offset)
+                self.assertEqual(base, list(expected))
